# Conflicts:
#	arches/app/models/graph.py

+++ tests/base_test.py
@@ -41,7 +41,8 @@ class ArchesTestCase(TestCase):
     @classmethod
     def deleteGraph(cls, root):
         def delete_children(node):
-            node.nodegroup.delete()
+            if node.nodegroup:
+                node.nodegroup.delete()
             for edge in models.Edge.objects.filter(rangenode=node):
                 edge.delete()
                 delete_children(edge.rangenode)
